+++ CHANGELOG.md
@@ -93,6 +93,9 @@ and this project adheres to
   need one of those, you're probably doing something wrong.
 - cosmwasm-std: Make `label` in `WasmMsg::Instantiate` non-optional to better
   match the Go/database format.
+- cosmwasm-std: `Binary::to_array` is now generic over the array length instead
+  of the output type. As a consequence the obsolete type `ByteArray` was
+  removed. The array length is not restricted to 0-64 anymore.
 - cosmwasm-vm: Avoid serialization of Modules in `InMemoryCache`, for
   performance. Also, remove `memory_limit` from `InstanceOptions`, and define it
   instead at `Cache` level (same memory limit for all cached instances).